It opens
added new datatypes, cleaned up the game launcher, make file works and made the "clean.h"

+++ bitmap.h
@@ -6,6 +6,6 @@ extern unsigned int alien_bitmap[8];
 
 
 unsigned int reverseBits(unsigned int num);
-void mirrorBitmap(const unsigned int *bitmap, unsigned int *mirroredBitmap, size_t size);
+void mirrorBitmap(const unsigned int *bitmap, unsigned int *mirroredBitmap, unsigned int size);
 
 #endif /* BITMAP_H */
